Enhanced circuit record and playback
The circuit recorder now records ciruits that exist in Tor but
were built before the recorder started listening for them. It
also records all circuits that exist in Tor but have not closed
yet (and so have not been written to the CSV) before the recorder
exits.

The circuit player now retries buidling a circuit for any circuit
that it tried to build but the circuit failed for some reason.