Better stream assignment
We now wait to assign streams to a circuit until we know that a
circuit was built successfully. This way, if a circuit times out,
we can retry on another circuit and if that one succeeds we will
then assign the stream. Hopefully this will allow more streams
to be successful.

Also, if a stream is detached from a circuit because the circuit
failed, we now will try to reattach the stream if we have a
circuit that will replace the failed one.

NOTICE: this commit introduces a SEGFAULT!
I am committing anyway for testing purposes, and will add a fix
in a future commit.