Refactor play mode
Circuits are now built reactively, after the first stream in a
new session is created. Future streams in the session will use an
already built circuit for that session.

If circuits fail or close, they will be re-created if they have
streams that need to finish.

If streams get detached, they will be attached to the latest
circuit available for that session. If one does not exit, it will
be built.

A session may have many circuits. We will now rotate from one
circuit to the next according to recorded circuit launch times.