Build scheduled circuits early
Added back in the feature that will build circuits earlier than
they were needed. The goal is to have the circuit built by the
time the tgen stream needs to use it.

In the latest code, if a tgen stream on a session is created
before the circuit is launched, then we will launch it at that
time (rather than waiting until the time the circuit was launched
in the recording).

In summary, we get the best of both worlds: try to build the
circuit so it is ready by the time we need it to be, and if for
some reason a stream is created before then, then we create the
circuit when we realize that it is needed.